refactor: fix depracated zod properties, created & used Roles enum and  and changed /redux/index.ts => /redux/store.ts
