+++ lib/providers/store-provider.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { persistor, store } from "@/lib/redux";
+import { persistor, store } from "@/lib/redux/store";
 import { Provider } from "react-redux";
 import { PersistGate } from "redux-persist/integration/react";
