+++ lib/redux/store.ts
@@ -9,7 +9,7 @@ const rootReducer = combineReducers({
 const persistConfig = {
   key: "root",
   storage,
-//   whitelist: ["auth"],
+  //   whitelist: ["auth"],
 };
 const persistedReducer = persistReducer(persistConfig, rootReducer);
