+++ lib/schemas/invite-user.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 export const inviteUserSchema = z.object({
-  email: z.string().nonempty("Email is required").email("Invalid email format"),
+  email: z.email("Invalid email format").nonempty("Email is required"),
   role: z.string().nonempty("Role is required"),
 });
 export type InviteUserInputs = z.infer<typeof inviteUserSchema>;
