+++ lib/types/index.ts
@@ -1,0 +1,6 @@
+export enum Roles {
+  LEARNER = "LEARNER",
+  MENTOR = "MENTOR",
+  MANAGER = "MANAGER",
+  ADMIN = "ADMIN",
+}
